% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ombrage.R
\name{calculer_ombrage_visualisation}
\alias{calculer_ombrage_visualisation}
\title{Visualiser l'ombrage d'une parcelle}
\usage{
calculer_ombrage_visualisation(
  resultats_ombrage,
  couche_heure = 1,
  titre = NULL
)
}
\arguments{
\item{resultats_ombrage}{Liste retournee par `calculer_ombrage()`}

\item{couche_heure}{Numero de la couche horaire a afficher (defaut: 1)}

\item{titre}{Titre optionnel pour le graphique}
}
\value{
NULL (affiche le graphique)
}
\description{
Cree une visualisation de l'ombrage calcule avec trois panneaux :
le MNE, l'ombrage moyen, et les heures d'ensoleillement.
}
\examples{
\dontrun{
resultats <- calculer_ombrage(champ, date = "2024-06-21")
visualiser_ombrage(resultats)
}

}
