% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{calculer_pente}
\alias{calculer_pente}
\title{Calculer la pente à partir d'un MNT}
\usage{
calculer_pente(mnt, filename = NULL)
}
\arguments{
\item{mnt}{Un objet `SpatRaster` représentant le MNT}

\item{filename}{Chemin de fichier pour sauvegarder le résultat (optionnel)}
}
\value{
Un objet `SpatRaster` contenant la pente en degrés
}
\description{
Calcule la pente en degrés à partir d'un modèle numérique de terrain (MNT).
}
\examples{
\dontrun{
mnt <- telecharger_lidar(champ)
pente <- calculer_pente(mnt)
}

}
