% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_vent.R
\name{fusionner_zones_vent}
\alias{fusionner_zones_vent}
\title{Fusionner et disjoindre les zones de vent par facteur H}
\usage{
fusionner_zones_vent(zones_sf)
}
\arguments{
\item{zones_sf}{Résultat de calculer_zones_vent_spline}
}
\value{
Un objet sf avec une ligne par facteur H, géométries fusionnées
}
\description{
Prend les zones brutes et retourne une couche où chaque facteur H est fusionné
et rendu disjoint des facteurs plus petits (effet de couronnes).
}
