% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labels_geomorphons}
\alias{labels_geomorphons}
\title{Obtenir les labels des géomorphons}
\usage{
labels_geomorphons()
}
\value{
Un vecteur caractère nommé
}
\description{
Retourne un vecteur nommé avec les labels des classes de géomorphons.
}
\examples{
labels <- labels_geomorphons()
print(labels)
}
