% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_vent.R, R/fetch_vent.R
\name{simuler_vitesse_vent}
\alias{simuler_vitesse_vent}
\title{Simuler la vitesse du vent}
\usage{
simuler_vitesse_vent(fetch, vitesse_ref = 5, coef_acceleration = 0.3)

simuler_vitesse_vent(fetch, vitesse_ref = 5, coef_acceleration = 0.3)
}
\arguments{
\item{fetch}{Résultat de calculer_fetch_vent()}

\item{vitesse_ref}{Vitesse de référence (m/s)}

\item{coef_acceleration}{Coefficient d'accélération après les arbres}

\item{result}{Résultat de calculer_distances_amont_aval()}

\item{coef_amont}{Effet ralentissement amont}

\item{coef_aval}{Effet accélération aval}
}
\value{
Raster de vitesse

Raster de vitesse
}
\description{
Simuler la vitesse du vent

Simuler la vitesse du vent avec effet de fetch
}
