% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_vent.R
\name{simuler_vitesse_vent_fetch}
\alias{simuler_vitesse_vent_fetch}
\title{Simuler la vitesse du vent avec effet de fetch}
\usage{
simuler_vitesse_vent_fetch(fetch, vitesse_ref = 5, coef_acceleration = 0.3)
}
\arguments{
\item{fetch}{Résultat de calculer_fetch_vent()}

\item{vitesse_ref}{Vitesse de référence (m/s)}

\item{coef_acceleration}{Coefficient d'accélération après les arbres}
}
\value{
Raster de vitesse
}
\description{
Simuler la vitesse du vent avec effet de fetch
}
