% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte_vent.R
\name{tracer_carte_vent}
\alias{tracer_carte_vent}
\title{Visualiser les distances amont/aval avec les flèches de vent}
\usage{
tracer_carte_vent(
  distances,
  type = c("les_deux", "amont", "aval", "total"),
  titre = NULL,
  n_fleches = 5,
  taille_fleches = 0.2
)
}
\arguments{
\item{distances}{Résultat de calculer_distances_vent()}

\item{type}{Type: "amont", "aval", "total" ou "les_deux"}

\item{titre}{Titre du graphique}

\item{n_fleches}{Nombre de flèches}

\item{taille_fleches}{Taille des flèches}
}
\value{
Un objet ggplot2
}
\description{
Visualiser les distances amont/aval avec les flèches de vent
}
\examples{
\dontrun{
carte <- tracer_carte_vent(distances, type = "amont")
print(carte)
}

}
