% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verifier_disponibilite_lidar}
\alias{verifier_disponibilite_lidar}
\title{Vérifier la disponibilité des données LiDAR}
\usage{
verifier_disponibilite_lidar(polygone, mne = FALSE)
}
\arguments{
\item{polygone}{Un objet `sf` représentant la zone d'intérêt ou un chemin de fichier}

\item{mne}{Logique. Si TRUE, vérifie la disponibilité du MNE. Sinon, du MNT.}
}
\value{
Un data.frame avec les années et nombre de tuiles disponibles
}
\description{
Vérifie si des données LiDAR sont disponibles pour une zone donnée
et retourne des informations sur les années disponibles.
}
\examples{
\dontrun{
info <- verifier_disponibilite_lidar(champ)
print(info)
}

}
