% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_ponctuel.R
\name{vider_cache_lidar}
\alias{vider_cache_lidar}
\title{Vider le cache LiDAR}
\usage{
vider_cache_lidar()
}
\value{
Message de confirmation
}
\description{
Supprime tous les fichiers LAZ téléchargés localement
}
\examples{
\dontrun{
vider_cache_lidar()
}
}
