% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_complet.R
\name{visualiser_covariables}
\alias{visualiser_covariables}
\title{Créer une visualisation synthétique des covariables extraites}
\usage{
visualiser_covariables(
  resultats,
  type = c("tous", "terrain", "ombrage", "vent", "arbres", "distances", "consolide")
)
}
\arguments{
\item{resultats}{Liste retournée par extraire_covariables_complet()}

\item{type}{Type de visualisation: "terrain", "ombrage", "vent", "arbres", "distances", "consolide", ou "tous"}
}
\value{
NULL (affiche les graphiques)
}
\description{
Créer une visualisation synthétique des covariables extraites
}
