% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_arbres.R
\name{visualiser_distance_arbres}
\alias{visualiser_distance_arbres}
\title{Visualiser la distance aux arbres}
\usage{
visualiser_distance_arbres(
  result,
  type = c("buffer", "points", "comparaison"),
  titre = NULL
)
}
\arguments{
\item{result}{Résultat de calculer_distance_arbres()}

\item{type}{"buffer", "points" ou "comparaison"}

\item{titre}{Titre optionnel}
}
\value{
ggplot2
}
\description{
Visualiser la distance aux arbres
}
