% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_vent.R
\name{visualiser_distances_vent}
\alias{visualiser_distances_vent}
\title{Visualiser les distances amont/aval}
\usage{
visualiser_distances_vent(
  result,
  type = c("comparaison", "amont", "aval", "totale")
)
}
\arguments{
\item{result}{Résultat de calculer_distances_amont_aval()}

\item{type}{"amont", "aval", "totale" ou "comparaison"}
}
\value{
ggplot2
}
\description{
Visualiser les distances amont/aval
}
