% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_vent.R
\name{visualiser_fetch}
\alias{visualiser_fetch}
\title{Visualiser le fetch de vent}
\usage{
visualiser_fetch(
  fetch,
  type = c("elliptique", "simple", "comparaison"),
  titre = NULL
)
}
\arguments{
\item{fetch}{Résultat de calculer_fetch_vent()}

\item{type}{"simple", "elliptique", ou "comparaison"}

\item{titre}{Titre optionnel}
}
\value{
ggplot2
}
\description{
Visualiser le fetch de vent
}
